Fix type error

+++ models/zoo.py
@@ -538,9 +538,9 @@ class ViTFree(nn.Module):
             self.prompt = None
 
         # classifier
-        self.num_centroids = prompt_param[1][1]
-        self.value_prototypes = torch.zeros((int(num_classes * self.num_centroids), 768))
-        self.prototype_std = torch.zeros((int(num_classes * self.num_centroids), 768))
+        self.num_centroids = int(prompt_param[1][1])
+        self.value_prototypes = torch.zeros((num_classes * self.num_centroids, 768))
+        self.prototype_std = torch.zeros((num_classes * self.num_centroids, 768))
 
         self._device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
